+++ Backend/src/main/java/com/realdolmen/repository/FlightRepository.java
@@ -7,7 +7,7 @@ import java.util.List;
 /**
  * Created by WVDAZ49 on 3/09/2016.
  */
-public interface FlightRepository extends Repository<Flight, Integer>{
+public interface FlightRepository extends Repository<Flight, Long>{
 
     List<Flight> getFlightsForAirport(String arrivalId, String departureId);
